split pods and hoist packages
