+++ pkg/hoist/.gitignore
