+++ pkg/hoist/fake_sv
