+++ pkg/hoist/hoist_executable.go
@@ -1,4 +1,4 @@
-package pods
+package hoist
 
 import (
 	"fmt"
