+++ pkg/hoist/hoist_executable_test.go
@@ -1,4 +1,4 @@
-package pods
+package hoist
 
 import (
 	"bytes"
