+++ pkg/hoist/hoist_launchable.go
@@ -1,4 +1,4 @@
-package pods
+package hoist
 
 import (
 	"archive/tar"
