+++ pkg/hoist/successful_scripts_test_hoist_launchable/installs/testLaunchable/bin/enable
