+++ pkg/hoist/test_helper.go
@@ -1,4 +1,4 @@
-package pods
+package hoist
 
 import (
 	"io/ioutil"
@@ -27,7 +27,7 @@ func FakeChpst() string {
 	return util.From(runtime.Caller(0)).ExpandPath("fake_chpst")
 }
 
-func fakeHoistLaunchableForDir(dirName string) *HoistLaunchable {
+func FakeHoistLaunchableForDir(dirName string) *HoistLaunchable {
 	tempDir, _ := ioutil.TempDir("", "fakeenv")
 	_, filename, _, _ := runtime.Caller(0)
 	launchableInstallDir := path.Join(path.Dir(filename), dirName)
